Working version of the subtitle scorer
We can now get the best match on subscene for a given video file name.
What remains to be done is to follow the link returned with that match,
download the subtitle ZIP archive, pipe it  an
unzipper then pipe the result to a file sharing the name of the video
file at the same path.

The video should at this point be playable with subtitles.